+++ packages/core/src/app/address/index.ts
@@ -1,6 +1,5 @@
 export { default as mapAddressToFormValues, AddressFormValues } from './mapAddressToFormValues';
 export { default as mapAddressFromFormValues } from './mapAddressFromFormValues';
-export { default as stripExtraFieldsFromAddress } from './stripExtraFieldsFromAddress';
 export { default as AddressForm } from './AddressForm';
 export { default as AddressFormModal } from './AddressFormModal';
 export { default as AddressSelect } from './AddressSelect';
@@ -11,7 +10,6 @@ export { default as isValidCustomerAddress } from './isValidCustomerAddress';
 export { default as isEqualAddress } from './isEqualAddress';
 export { default as setDefaultAddress } from './setDefaultAddress';
 export { default as getAddressWithLabel } from './getAddressWithLabel';
-export { default as getAddressWithCustomerExtraFields } from './getAddressWithCustomerExtraFields';
 export { reorderAddressFormFields } from './reorderAddressFormFields';
 export {
     default as getAddressFormFieldsValidationSchema,