+++ packages/core/src/app/billing/Billing.tsx
@@ -1,5 +1,5 @@
 import type { CheckoutSelectors, FormField } from '@bigcommerce/checkout-sdk';
-import React, { type ReactElement, useEffect, useMemo, useState } from 'react';
+import React, { type ReactElement, useCallback, useEffect, useState } from 'react';
 
 import { useCapabilities, useCheckout } from '@bigcommerce/checkout/contexts';
 import { TranslatedString } from '@bigcommerce/checkout/locale';
@@ -8,7 +8,6 @@ import { B2BSessionStorage } from '@bigcommerce/checkout/utility';
 
 import {
     AddressType,
-    getAddressWithCustomerExtraFields,
     isEqualAddress,
     mapAddressFromFormValues,
     setDefaultAddress,
@@ -41,55 +40,63 @@ const getFieldsWithExtraFields = (
 };
 
 const Billing = ({ navigateNextStep, onReady, onUnhandledError }: BillingProps): ReactElement => {
-    const { checkoutService, checkoutState } = useCheckout();
+    const {
+        selectedState: {
+            checkout,
+            config,
+            cart,
+            customer,
+            isLoadingBillingCountries,
+            getBillingAddressFields,
+            getAddressExtraFields,
+        },
+        // using getBillingAddress function to guarantee latest state inside of async function in useEffect
+        checkoutState: {
+            data: { getBillingAddress },
+        },
+        checkoutService,
+    } = useCheckout(({ data, statuses }) => ({
+        checkout: data.getCheckout(),
+        config: data.getConfig(),
+        cart: data.getCart(),
+        customer: data.getCustomer(),
+        billingAddress: data.getBillingAddress(),
+        isLoadingBillingCountries: statuses.isLoadingBillingCountries(),
+        getBillingAddressFields: data.getBillingAddressFields,
+        getAddressExtraFields: data.getAddressExtraFields,
+    }));
     const {
         userJourney: { hasAddressExtraFields, hasCompanyAddressBook },
         billing: { restrictManualAddressEntry },
     } = useCapabilities();
-
-    const {
-        data: {
-            getCheckout,
-            getConfig,
-            getCart,
-            getCustomer,
-            getBillingAddress,
-            getBillingAddressFields,
-            getAddressExtraFields,
-        },
-        statuses: { isLoadingBillingCountries },
-    } = checkoutState;
-    const config = getConfig();
-    const customer = getCustomer();
-    const checkout = getCheckout();
-    const cart = getCart();
 
     if (!config || !customer || !checkout || !cart) {
         throw new Error('Unable to access checkout data');
     }
 
     const [isApplyingDefaultAddress, setIsApplyingDefaultAddress] = useState(true);
-    const isInitializing = isLoadingBillingCountries() || isApplyingDefaultAddress;
+    const isInitializing = isLoadingBillingCountries || isApplyingDefaultAddress;
 
     // Below constants are for <BillingForm />'s HOC props
     const customerMessage = checkout.customerMessage;
     const methodId = getBillingMethodId(checkout);
-    const rawBillingAddress = getBillingAddress();
-    const billingAddress = useMemo(
-        () =>
-            getAddressWithCustomerExtraFields(
-                rawBillingAddress,
-                customer.addresses,
+    const billingAddress = getBillingAddress();
+
+    const getFields = useCallback(
+        (countryCode?: string) =>
+            getFieldsWithExtraFields(
+                getBillingAddressFields,
                 hasAddressExtraFields,
+                getAddressExtraFields,
+                countryCode,
             ),
-        [hasAddressExtraFields, rawBillingAddress, customer.addresses],
+        [getBillingAddressFields, hasAddressExtraFields, getAddressExtraFields],
     );
+
     const handleSubmit = async ({
         orderComment,
         ...addressValues
     }: BillingFormValues): Promise<void> => {
-        const updateAddress = checkoutService.updateBillingAddress;
-        const updateCheckout = checkoutService.updateCheckout;
         const billingAddress = getBillingAddress();
         const promises: Array<Promise<CheckoutSelectors>> = [];
         const address = mapAddressFromFormValues(
@@ -98,11 +105,11 @@ const Billing = ({ navigateNextStep, onReady, onUnhandledError }: BillingProps):
         );
 
         if (address && !isEqualAddress(address, billingAddress)) {
-            promises.push(updateAddress(address));
+            promises.push(checkoutService.updateBillingAddress(address));
         }
 
         if (customerMessage !== orderComment) {
-            promises.push(updateCheckout({ customerMessage: orderComment }));
+            promises.push(checkoutService.updateCheckout({ customerMessage: orderComment }));
         }
 
         try {
@@ -166,14 +173,7 @@ const Billing = ({ navigateNextStep, onReady, onUnhandledError }: BillingProps):
                 <BillingForm
                     billingAddress={billingAddress}
                     customerMessage={customerMessage}
-                    getFields={(countryCode?: string) =>
-                        getFieldsWithExtraFields(
-                            getBillingAddressFields,
-                            hasAddressExtraFields,
-                            getAddressExtraFields,
-                            countryCode,
-                        )
-                    }
+                    getFields={getFields}
                     methodId={methodId}
                     navigateNextStep={navigateNextStep}
                     onSubmit={handleSubmit}