+++ packages/core/src/app/checkout/components/CheckoutHeader.tsx
@@ -28,14 +28,17 @@ export const CheckoutHeader: React.FC<CheckoutHeaderProps> = ({
     onUnhandledError,
     onWalletButtonClick,
 }) => {
-    const { checkoutState } = useCheckout();
+    const {
+        selectedState: { checkout, config, flashMessages },
+    } = useCheckout(({ data }) => ({
+        checkout: data.getCheckout(),
+        config: data.getConfig(),
+        flashMessages: data.getFlashMessages('warning'),
+    }));
     const { extensionState } = useExtensions();
 
-    const { data } = checkoutState;
+    const { promotions = EMPTY_ARRAY } = checkout || {};
 
-    const { promotions = EMPTY_ARRAY } = data.getCheckout() || {};
-
-    const config = data.getConfig();
     const isShowingWalletButtonsOnTop = Boolean(
         config?.checkoutSettings?.checkoutUserExperienceSettings?.walletButtonsOnTop,
     );
@@ -46,9 +49,7 @@ export const CheckoutHeader: React.FC<CheckoutHeaderProps> = ({
 
     return (
         <>
-            <BackorderQuantitiesChangedBanner
-                message={data.getFlashMessages('warning')?.[0]?.message}
-            />
+            <BackorderQuantitiesChangedBanner message={flashMessages?.[0]?.message} />
             <LoadingNotification isLoading={extensionState.isShowingLoadingIndicator} />
             <PromotionBannerList promotions={promotions} />
             {isShowingWalletButtonsOnTop && buttonConfigs?.length > 0 && (