+++ packages/core/src/app/payment/PaymentSubmitButton.test.tsx
@@ -66,14 +66,6 @@ describe('PaymentSubmitButton', () => {
         ).toBeInTheDocument();
     });
 
-    it('renders button with special label for Barclaycard', () => {
-        render(<PaymentSubmitButtonTest methodGateway="barclaycard" />);
-
-        expect(
-            screen.getByText(languageService.translate('payment.barclaycard_continue_action')),
-        ).toBeInTheDocument();
-    });
-
     it('renders button with special label for Visa Checkout provided by Braintree', () => {
         render(<PaymentSubmitButtonTest methodType="visa-checkout" />);