+++ packages/error-handling-utils/src/InstrumentDeclinedError/InstrumentDeclinedError.ts
@@ -1,0 +1,13 @@
+import { CustomError } from '@bigcommerce/checkout/payment-integration-api';
+import { getLanguageService } from '@bigcommerce/checkout/locale';
+
+export default class InstrumentDeclinedError extends CustomError {
+    constructor() {
+        super({
+            message: getLanguageService().translate('payment.errors.instrument_declined'),
+            name: 'INSTRUMENT_DECLINED_ERROR',
+        });
+
+        Object.setPrototypeOf(this, InstrumentDeclinedError.prototype);
+    }
+}