+++ packages/ui/src/form/DynamicFormField/DynamicFormFieldSelector.tsx
@@ -23,6 +23,7 @@ interface DynamicFormFieldSelectorProps {
     maxLength?: FormFieldType['maxLength'];
     isFloatingLabelEnabled?: boolean;
     isNewPhoneValidationExperimentEnabled: boolean;
+    required?: boolean;
     selectedCountry?: string;
     onChange?(value: string | string[]): void;
 }
@@ -42,6 +43,7 @@ export const DynamicFormFieldSelector: FunctionComponent<DynamicFormFieldSelecto
         maxLength,
         isFloatingLabelEnabled,
         isNewPhoneValidationExperimentEnabled,
+        required,
         selectedCountry,
         onChange,
     }) => {
@@ -101,6 +103,7 @@ export const DynamicFormFieldSelector: FunctionComponent<DynamicFormFieldSelecto
                     maxLength={maxLength || undefined}
                     name={name}
                     onChange={onChange}
+                    required={required}
                     selectedCountry={selectedCountry}
                 />
             );